+++ src/main/kotlin/ru/spbstu/competition/protocol/Protocol.kt
@@ -3,7 +3,7 @@ package ru.spbstu.competition.protocol
 import ru.spbstu.competition.protocol.data.*
 
 class Protocol(val connection: ServerConnection) {
-    constructor(url: String, port: Short): this(ServerConnection(url, port))
+    constructor(url: String, port: Int): this(ServerConnection(url, port))
 
     // После прописывания всех наших типов данных в Messages.kt
     // и грамотной реализации операций в ServerConnection
